chore: rename EyebrowLabel component file
