+++ src/pages/community/components/Channels.tsx
@@ -1,7 +1,7 @@
 import {Github,MessageCircle,Linkedin,Twitter,ArrowUpRight} from "lucide-react";
 import {  SOCIAL_PLATFORMS } from "@/constants";
 import { ScrollAnimatedItem } from "@/components/UI/ScrollAnimatedItem";
-import EyebrowLabel from "@/components/UI/EyebrowLable";
+import EyebrowLabel from "@/components/UI/EyebrowLabel";
